requirements.txt
fastapi
uvicorn
streamlit
scikit-learn
pandas
numpy
joblib
pyyaml
requests
groq
python-dotenv
